split directoryentry made classes noncopyable, fixed a few unintended copies
