+++ src/main.cpp
@@ -48,6 +48,7 @@ along with Mod Organizer.  If not, see <http://www.gnu.org/licenses/>.
 #include "organizercore.h"
 #include "env.h"
 #include "envmodule.h"
+#include "util.h"
 
 #include <eh.h>
 #include <windows_error.h>
