+++ src/pch.h
@@ -5,14 +5,17 @@
 #include <cstdarg>
 #include <cstdint>
 #include <ctime>
+#include <deque>
 #include <exception>
 #include <functional>
+#include <fstream>
 #include <iomanip>
 #include <iostream>
 #include <limits.h>
 #include <list>
 #include <map>
 #include <memory>
+#include <mutex>
 #include <optional>
 #include <regex>
 #include <set>
@@ -21,6 +24,7 @@
 #include <stdexcept>
 #include <string.h>
 #include <string>
+#include <thread>
 #include <tuple>
 #include <unordered_set>
 #include <utility>
